Protocol Changes + Tests

+++ src/main/java/me/outspending/position/Location.java
@@ -1,6 +1,6 @@
 package me.outspending.position;
 
-public record Location(long x, long y, long z) {
+public record Location(int x, int y, int z) {
     public static Location ZERO = new Location(0, 0, 0);
     public static Location ONE = new Location(1, 1, 1);
 }
